industrial_copilot/backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
pgvector
pymupdf
easyocr
camelot-py[cv]
ultralytics
huggingface_hub
openai
langchain
langchain-openai
langchain-core
langgraph
influxdb-client
tensorflow
scikit-learn
pandas
numpy
python-multipart
python-dotenv
cloudinary
